requirements.txt
httpx
python-telegram-bot[all]
python-decouple
python-dotenv
sqlalchemy
cachetools